BCE structure established
