+++ articles-java-jee/src/main/java/com/ibm/cns/articles/boundary/JAXRSConfiguration.java
@@ -1,0 +1,8 @@
+package com.ibm.cns.articles.boundary;
+
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+@ApplicationPath("/")
+public class JAXRSConfiguration extends Application {
+}
