+++ articles-java-jee/src/main/java/com/ibm/cns/articles/control/ArticleDao.java
@@ -1,12 +1,11 @@
 package com.ibm.cns.articles.control;
 
-import com.ibm.cns.articles.entity.ArticleEntity;
-import com.ibm.cns.articles.entity.NoConnectivity;
+import com.ibm.cns.articles.entity.Article;
 import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import javax.annotation.Resource;
 import javax.enterprise.context.ApplicationScoped;
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import javax.transaction.UserTransaction;
 
 @ApplicationScoped
@@ -18,7 +17,7 @@ public class ArticleDao {
     @Resource
     UserTransaction utx;
 
-    public void createArticle(ArticleEntity article) throws NoConnectivity {
+    public void createArticle(Article article) throws NoConnectivity {
         try {
             utx.begin();
             em.persist(article);
@@ -29,30 +28,31 @@ public class ArticleDao {
         }
     }
 
-    public ArticleEntity readArticle(int articleId) {
-        return em.find(ArticleEntity.class, articleId);
+    public Article readArticle(int articleId) {
+        return em.find(Article.class, articleId);
     }
 
-    public void updateArticle(ArticleEntity article) {
+    public void updateArticle(Article article) {
         em.merge(article);
     }
 
-    public void deleteArticle(ArticleEntity article) {
+    public void deleteArticle(Article article) {
         em.remove(article);
     }
 
-    public List<ArticleEntity> readAllArticles() {
-        return em.createNamedQuery("Article.findAll", ArticleEntity.class).getResultList();
+    public List<Article> readAllArticles() {
+        return em.createNamedQuery("Article.findAll", Article.class).getResultList();
     }
 
-    public List<ArticleEntity> find(int articleId) {
-        return em.createNamedQuery("Article.findArticle", ArticleEntity.class).setParameter("id", articleId)
+    public List<Article> find(int articleId) {
+        return em.createNamedQuery("Article.findArticle", Article.class).
+                setParameter("id", articleId)
                 .getResultList();
 
     }
 
-    public List<ArticleEntity> findArticle(String title, String url, String author) {
-        return em.createNamedQuery("Article.findArticle", ArticleEntity.class).setParameter("title", title)
+    public List<Article> findArticle(String title, String url, String author) {
+        return em.createNamedQuery("Article.findArticle", Article.class).setParameter("title", title)
                 .setParameter("url", url).setParameter("author", author).getResultList();
     }
 }
