+++ articles-java-jee/src/main/java/com/ibm/cns/articles/control/DataAccess.java
@@ -1,9 +1,9 @@
-package com.ibm.cns.articles.entity;
+package com.ibm.cns.articles.control;
 
 import java.util.List;
 
 import com.ibm.cns.articles.entity.Article;
-import com.ibm.cns.articles.entity.ArticleDoesNotExist;
+import com.ibm.cns.articles.entity.Article;
 
 public interface DataAccess {
 	public Article addArticle(Article article) throws NoConnectivity;
